engine: extract esm/utilities/color_picker.js
