+++ engine/esm/index.js
@@ -219,6 +219,7 @@ export { Pointing } from "./pointing.js";
 export { Tour } from "./tour.js";
 export { VideoOutputType } from "./video_output_type.js";
 export { VizLayer } from "./viz_layer.js";
+export { ColorPicker } from "./utilities/color_picker.js";
 export { Histogram } from "./utilities/histogram.js";
 
 export { Layer, DomainValue, AltUnits, FadeType } from "./layers/layer.js";
@@ -318,7 +319,6 @@ export {
     UndoStep,
     UndoTourSlidelistChange,
     UndoTourPropertiesChange,
-    ColorPicker,
     Dialog,
     FolderDownloadAction,
     Wtml,
